pyproject.toml
[project]
name = "content-curator"
version = "0.1.0"
description = "Automatically generate readable, well-sourced blog articles from tech social media"
authors = [{name = "Your Name", email = "your.email@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.14"  # Python 3.14+ for free-threading support

dependencies = [
    # Core dependencies
    "pydantic>=2.5.0",
    "httpx>=0.25.0",
    "rich>=13.7.0",
    # AI & Content Generation
    "openai>=1.3.0",
    # Social Media APIs
    "atproto>=0.0.20", # Bluesky
    "mastodon.py>=1.8.0",
    "praw>=7.7.0", # Reddit (optional)
    # Content Processing
    "python-frontmatter>=1.1.0",
    "python-dotenv>=1.0.0",
    "tenacity>=9.1.2",
    "requests>=2.32.5",
    "beautifulsoup4>=4.14.2",
    # URL normalization (lightweight, RFC-friendly)
    "url-normalize>=1.4.3",
    # Image processing for icon generation
    "pillow>=10.0.0",
    # Markdown formatting and linting
    "mdformat>=0.7.17",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
    "types-requests",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py314"
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by formatter
    "B008",  # do not perform function calls in argument defaults
    "E402",  # module level import not at top of file (used for conditional imports)
    "E741",  # ambiguous variable name (single letter vars in existing code)
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.14"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "mastodon.*",
    "praw.*",
    "atproto.*",
]
ignore_missing_imports = true
